requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
exceptiongroup==1.3.1
fastapi==0.128.8
flatbuffers==25.12.19
gast==0.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.16.0
h5py==3.14.0
httplib2==0.31.2
idna==3.11
importlib_metadata==8.7.1
joblib==1.5.3
keras==2.15.0
libclang==18.1.1
Markdown==3.9
MarkupSafe==3.0.3
ml-dtypes==0.2.0
numpy==1.26.4
oauthlib==3.3.1
opt_einsum==3.4.0
opencv-python==4.10.0.84
packaging==26.0
pillow==11.3.0
proto-plus==1.27.1
protobuf==4.21.12
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
starlette==0.49.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.39.0
Werkzeug==3.1.5
wrapt==1.14.2
zipp==3.23.0